dataset_generation/requirements.txt
openai>=1.0.0
transformers>=4.35.0
torch>=2.0.0
pyyaml>=6.0
tqdm>=4.65.0
tenacity>=8.2.0

